组织架构功能实现

+++ morefocus-base/src/main/java/cn/morefocus/base/module/support/serialnumber/domain/bo/SerialNumberAddFrom.java
@@ -11,7 +11,7 @@ import javax.validation.constraints.NotNull;
 public class SerialNumberAddFrom {
 
     @NotNull(message = "ID不能为空")
-    private Integer serialNumberId;
+    private Long serialNumberId;
 
     /**
      * 编码前缀
